Add optional I²C-Display compatability
There are many 16x2-Displays with a piggybagged I²C-Portexpander out
there. These are easily supported by the LiquidCrystal_I2C-Library.

This commit modifies the output field to be a byte instead of bool, and
adds option output=2 for the I²C-Display. To include it, #define I2CDISP
is also needed.